% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{load, check and plot input data for use in breakpoint detection}
\usage{
check_data(dat, sex = FALSE, showPlot = TRUE)
}
\arguments{
\item{dat}{data.frame with columns year, age, length, sex (optional)}

\item{sex}{logical. does your data frame include a sex column?}

\item{showPlot}{logical. do you want to visualize your input data?}
}
\value{
plots
}
\description{
load, check and plot input data for use in breakpoint detection
}
\examples{
\dontrun{
read_rep(fn = 'inst/example_data/goasr.rep')
}
}
