% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Deriv.R
\name{check_Deriv}
\alias{check_Deriv}
\title{Check whether the CI of the derivative includes zero or not}
\usage{
check_Deriv(x, d, upper, lower, eval = 0, crit.eval)
}
\arguments{
\item{x}{vector of values over which derivatives were evaluated}

\item{d}{vector output of {Deriv()}}

\item{upper}{upper confidence interval; output of {confint_Deriv()}}

\item{lower}{lower confidence interval; output of {confint_Deriv()}}

\item{eval}{the value that you would like to examine. default 0}
}
\value{
vector of confidence intervals of same length as object.
}
\description{
Check whether the CI of the derivative includes zero or not
}
