% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_Deriv.R
\name{confint_Deriv}
\alias{confint_Deriv}
\title{Calculate confidence interval of derivative, adapted from
\href{https://gist.githubusercontent.com/gavinsimpson/e73f011fdaaab4bb5a30/raw/82118ee30c9ef1254795d2ec6d356a664cc138ab/Deriv.R}{Gavin Simpson}.}
\usage{
confint_Deriv(object, term, alpha = 0.05, ...)
}
\arguments{
\item{object}{vector output of {Deriv()}}

\item{term}{string smooth name, i.e. "year". must match values in mod and newdata}

\item{alpha}{confidence threshold. default 0.05 (for 95% CI)}
}
\value{
vector of confidence intervals of same length as object.
}
\description{
Calculate confidence interval of derivative, adapted from
\href{https://gist.githubusercontent.com/gavinsimpson/e73f011fdaaab4bb5a30/raw/82118ee30c9ef1254795d2ec6d356a664cc138ab/Deriv.R}{Gavin Simpson}.
}
