% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deriv.R
\name{Deriv}
\alias{Deriv}
\title{Evaluate first derivative of GAM smooth(s), adapted from
\href{https://gist.githubusercontent.com/gavinsimpson/e73f011fdaaab4bb5a30/raw/82118ee30c9ef1254795d2ec6d356a664cc138ab/Deriv.R}{Gavin Simpson}.}
\usage{
Deriv(mod, n = 200, eps = 1e-04, newdata, term)
}
\arguments{
\item{mod}{the output of {mgcv::gam()}}

\item{n}{number of intervals over which to evaluate each smooth. default 200.}

\item{eps}{tolerance threshold. default 1e-4.}

\item{newdata}{optional; data.frame of new smooth parameters to evaluate}

\item{term}{string smooth name, i.e. "year". must match values in mod and newdata}
}
\value{
a vector of derivative values of length n
}
\description{
Evaluate first derivative of GAM smooth(s), adapted from
\href{https://gist.githubusercontent.com/gavinsimpson/e73f011fdaaab4bb5a30/raw/82118ee30c9ef1254795d2ec6d356a664cc138ab/Deriv.R}{Gavin Simpson}.
}
