% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Breaks.R
\name{get_Breaks}
\alias{get_Breaks}
\title{Wrapper function to to fit gams & evaluate first derivatives}
\usage{
get_Breaks(dat, ages_to_use = c(5, 10), sex = FALSE, axes = 0, showPlot = TRUE)
}
\arguments{
\item{dat}{data.frame with columns year, age, length, lat, long, sex (optional)}

\item{ages_to_use}{optional. vector of age(s) to evaluate for breakpoints.}

\item{sex}{logical.}

\item{axes}{do you want to evaluate axes for space only (0, default), time only (1), or both simultaneously (2)?}

\item{showPlot}{logical. do you want to see the detected break(s) on a map? Applicable only for axes options 0 or 2.}
}
\value{
matrix of detected breakpoints and plots
}
\description{
Wrapper function to to fit gams & evaluate first derivatives
}
\examples{
 \dontrun{ data(simulated_data) ;get_Breaks(dat = simulated_data, ages_to_use = c(15:20), axes = 0)
}
}
