% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthbreaks-package.R, R/src-package.R
\docType{package}
\name{growthbreaks-package}
\alias{growthbreaks}
\alias{growthbreaks-package}
\title{growthbreaks: Spatio-temporal Growth Anomaly Detection}
\description{
Implements the GAM-based breakpoint detection method for length-at-age observations as described Kapur et. al. 2019.

Implements the GAM-based breakpoint detection method for length-at-age observations as described Kapur et. al. 2019.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://afsc-assessments.github.io/growthbreaks/}
}


Useful links:
\itemize{
  \item \url{https://afsc-assessments.github.io/growthbreaks/}
}

}
\keyword{internal}
