% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_Growth.R
\name{refit_Growth}
\alias{refit_Growth}
\title{Function to re-fit growth data at  putative breaks and return estimates for validation}
\usage{
refit_Growth()
}
\arguments{
\item{x}{vector of values over which derivatives were evaluated}

\item{d}{vector output of {Deriv()}}

\item{upper}{upper confidence interval; output of {confint_Deriv()}}

\item{lower}{lower confidence interval; output of {confint_Deriv()}}

\item{eval}{the value that you would like to examine. default 0}
}
\value{
Von B growth parameters at input breakpoints; plots with uncertainty of growth curves
}
\description{
Function to re-fit growth data at  putative breaks and return estimates for validation
}
