% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Breaks.R
\name{get_Breaks}
\alias{get_Breaks}
\title{Wrapper function to to fit gams & evaluate first derivatives}
\usage{
get_Breaks(dat, axes = "both")
}
\arguments{
\item{x}{vector of values over which derivatives were evaluated}

\item{d}{vector output of {Deriv()}}

\item{upper}{upper confidence interval; output of {confint_Deriv()}}

\item{lower}{lower confidence interval; output of {confint_Deriv()}}

\item{eval}{the value that you would like to examine. default 0}
}
\value{
matrix of detected breakpoints and plots
}
\description{
Wrapper function to to fit gams & evaluate first derivatives
}
